Added results without optimisation of samples with known linear objective
Noticed that Grammacy's code was optimised in the case of a known linear
objecctive; when sampling points for optimising the acquisition function
he was restricting the sampling space to only be the region in which the
sum of inputs was less than the best objective value seen so far (since
the objective is linear).

Also ran the code with 6000 initial samples for optimisation, in line
with the Trieste experiments.