Make final changes for fair comparison with my approach
Previously the number of random samples used to optimise the acqusisiton
function differed between mine and Gramacy's code. In the case that
<10% of randomly sampled points had EI>0, he would resample 10 times
the number of random points and calculate EI again, since EI can become 0
in a lot of the search space in later iterations.

I have removed this, so that we just sample the same number of points all
the time in his implementation and my implementation for a fair comparison.
In line with the Trieste recommendations, this will be min(5000, 1000*d),
where d is the dimensionality og the search space.